Restructure and update as needed
