+++ requirements.txt
@@ -1,0 +1,2 @@
+PyQt5==5.15.9
+pyserial==3.5
