+++ serial_reader/__init__.py
@@ -1,0 +1,2 @@
+# Optionally, you can expose SerialReaderThread directly in the package namespace
+from .SerialReaderThread import SerialReaderThread
