+++ settings.ini
@@ -1,3 +1,3 @@
 [user_settings]
-selected_port = 3
+selected_port = 0
