+++ views/__init__.py
@@ -1,0 +1,3 @@
+# Optionally, you can expose SerialReaderThread directly in the package namespace
+from .Data_View_Window import DataViewWindow
+from .Main_Window import MainWindow
